\name{ade4TkGUI-internal}
\alias{chooseaxes}
\alias{choosechull}
\alias{choosecol}
\alias{choosecw}
\alias{choosedf}
\alias{choosedist}
\alias{choosedudi}
\alias{choosedudirc}
\alias{chooseduditest}
\alias{choosefac}
\alias{choosefac2}
\alias{choosegraph}
\alias{chooselab}
\alias{chooseneig}
\alias{choosepackage}
\alias{choosepm}
\alias{chooserw}
\alias{choosesp}
\alias{chooseval}
\alias{choosewt}
\alias{dialog.between}
\alias{dialog.cca}
\alias{dialog.coinertia}
\alias{dialog.discrimin}
\alias{dialog.dpcoa}
\alias{dialog.dudi.acm}
\alias{dialog.dudi.coa}
\alias{dialog.dudi.dec}
\alias{dialog.dudi.display}
\alias{dialog.dudi.fca}
\alias{dialog.dudi.fpca}
\alias{dialog.dudi.mix}
\alias{dialog.dudi.nsc}
\alias{dialog.dudi.pca}
\alias{dialog.dudi.pco}
\alias{dialog.MCTests}
\alias{dialog.pcaiv}
\alias{dialog.pcaivortho}
\alias{dialog.s.arrow}
\alias{dialog.s.chull}
\alias{dialog.s.class}
\alias{dialog.s.corcircle}
\alias{dialog.s.label}
\alias{dialog.s.match}
\alias{dialog.s.value}
\alias{dialog.within}
\alias{displaytable}
\alias{dudisp}
\alias{exploregraph}
\alias{newGr}
\alias{resetgraph}
\alias{outgraph}
\alias{readtable}
\alias{selectGr}
\alias{askQuit}
\alias{editdf}
\alias{rewriteHistory}
\title{Internal functions for the ade4TkGUI package.}
\description{
Internal functions for the ade4TkGUI package
}
\details{
These are not to be called by the user.
}
\keyword{multivariate}
